Refactor [#67]: Rename MemoryTest to DefaultMemoryTest (#71)
